Conver MPII annotations from .mat to .json Version 2
The former version was the DeepHighResolution Net version.
This is one put together through a StackOverflow question&answer.
And this is still not what bmartacho used based on mpii_data.py#L85,
where an isValidation field is used.
So that part will be refactored, because I can't find any more json version of the MPII annotations.